style : 모호한 API 이름을 규칙에 맞춰서 변경
DESC
----
- 특별한 원칙 없이 정해진 frontend api 이름을 통일감 있게 변경
