+++ packages/frontend/src/api/confirmJoinUser.test.tsx
@@ -1,13 +1,16 @@
-import { CreateUserDTO, User } from '@my-task/common';
+import { CreateUserConfirmDTO, User } from '@my-task/common';
 import { QueryClient, QueryClientProvider, UseMutationResult } from '@tanstack/react-query';
 import { RenderHookResult, renderHook, waitFor } from '@testing-library/react';
 import { describe, expectTypeOf } from 'vitest';
-import postJoin from '~/api/postJoin';
 import { BE_ORIGIN } from '~/constants';
 import { server } from '~/mock';
+import confirmJoinUser from './confirmJoinUser';
 
-describe('postJoin', () => {
-  let renderedHook: RenderHookResult<UseMutationResult<any, any, CreateUserDTO, any>, unknown>;
+describe('requestJoinUser', () => {
+  let renderedHook: RenderHookResult<
+    UseMutationResult<any, any, CreateUserConfirmDTO, any>,
+    unknown
+  >;
 
   const testQueryClient = new QueryClient({
     defaultOptions: {
@@ -32,7 +35,7 @@ describe('postJoin', () => {
   afterAll(() => server.close());
 
   beforeEach(() => {
-    renderedHook = renderHook(() => postJoin({}), { wrapper });
+    renderedHook = renderHook(() => confirmJoinUser({}), { wrapper });
   });
 
   afterEach(() => {
@@ -40,18 +43,19 @@ describe('postJoin', () => {
   });
 
   it('should work', async () => {
-    const dto: CreateUserDTO = { email: 'test@email.com' };
+    const dto: CreateUserConfirmDTO = { uuid: '993ae2a1-2554-404c-8a86-660b5ee7fedd' };
     renderedHook.result.current.mutate(dto);
     await waitFor(() => expect(renderedHook.result.current.isSuccess).toEqual(true));
 
     const data = renderedHook.result.current.data;
     expectTypeOf(data).toMatchTypeOf<User>();
-    expect(data.email).toEqual(dto.email);
+    expect(data.email).toEqual('success@example.com');
   });
 
   describe('should fail with', () => {
     it('invalid email', async () => {
-      renderedHook.result.current.mutate({ email: 'invalid@email' });
+      const dto: CreateUserConfirmDTO = { uuid: '6aa6ee8e-a4f8-49f6-817f-1c9342aae29e' };
+      renderedHook.result.current.mutate(dto);
       await waitFor(() => expect(renderedHook.result.current.isError).toEqual(true));
     });
   });
