+++ packages/frontend/src/api/requestJoinUser.ts
@@ -3,10 +3,10 @@ import { useMutation } from '@tanstack/react-query';
 import { MutationOptions } from '~/types';
 import fetchCore from './fetchCore';
 
-const postJoin = (options: MutationOptions<CreateUserDTO, CreateUserDTO>) =>
+const requestJoinUser = (options: MutationOptions<CreateUserDTO, CreateUserDTO>) =>
   useMutation({
     mutationFn: (body) => fetchCore('/auth', { method: 'POST', body }),
     ...options,
   });
 
-export default postJoin;
+export default requestJoinUser;
